clean up useids memory
